refactor: unify differential test crates
