--- ckb-vm-differential-guest/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ckb-vm-differential-guest"
-description = "Guest-side helpers for ckb-vm differential testing"
-version = "1.1.1"
-license = "MIT"
-edition = "2024"
-authors = ["Nervos Core Dev <dev@nervos.org>"]
-
-[dependencies]
-ckb-vm-differential-protocol = { path = "../ckb-vm-differential-protocol", version = "1.1.1" }
-serde = { version = "1.0", default-features = false, features = ["derive", "alloc"] }
-postcard = { version = "1.0", default-features = false, features = ["alloc"] }
