--- ckb-vm-differential-host/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "ckb-vm-differential-host"
-description = "Host-side harness for differential testing of cryptography libraries ported to ckb-vm"
-version.workspace = true
-license.workspace = true
-edition.workspace = true
-authors.workspace = true
-
-[dependencies]
-ckb-vm-differential-protocol = { path = "../ckb-vm-differential-protocol", version = "1.1.1" }
-ckb-vm.workspace = true
-serde = { workspace = true, features = ["derive"] }
-postcard = { version = "1.0", features = ["use-std"] }
-proptest = "1.5"
-thiserror = "1.0"
