--- ckb-vm-differential-protocol/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "ckb-vm-differential-protocol"
-description = "Shared wire-protocol definitions for ckb-vm differential testing"
-version.workspace = true
-license.workspace = true
-edition.workspace = true
-authors.workspace = true
-
-[dependencies]
