--- ckb-vm-differential/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "ckb-vm-differential"
-description = "Facade crate for ckb-vm differential testing: one `harness!` call per library under test"
-version.workspace = true
-license.workspace = true
-edition.workspace = true
-authors.workspace = true
-
-[dependencies]
-ckb-vm-differential-protocol = { path = "../ckb-vm-differential-protocol", version = "1.1.1" }
-
-[target.'cfg(target_arch = "riscv64")'.dependencies]
-ckb-vm-differential-guest = { path = "../ckb-vm-differential-guest", version = "1.1.1" }
-
-[target.'cfg(not(target_arch = "riscv64"))'.dependencies]
-ckb-vm-differential-host = { path = "../ckb-vm-differential-host", version = "1.1.1" }
