+++ ckb-vm-test-suite/programs/contracts/sha256_differential/Cargo.toml
@@ -22,7 +22,7 @@ name = "proptest"
 path = "tests/proptest.rs"
 
 [dependencies]
-ckb-vm-differential = { path = "../../../../ckb-vm-differential", version = "1.1.1" }
+ckb-vm-differential-test = { path = "../../../../ckb-vm-differential-test", version = "1.1.1" }
 
 [target.'cfg(not(target_arch = "riscv64"))'.dependencies]
 sha2 = "0.10"
@@ -30,4 +30,3 @@ proptest = "1.5"
 
 [target.'cfg(target_arch = "riscv64")'.dependencies]
 ckb-opt-sha256 = { path = "../../../../opt-lib/sha256/rust", version = "0.1.0" }
-ckb-std = "0.17"
